Drop legacy unit discovery document
Remove the extra nest-native discovery payload and rely on Home
Assistant device discovery as the only discovery contract for this
slice.

Update startup publishing and tests to expect only Home Assistant
 discovery and availability messages.
