+++ internal/mqtt/startup_test.go
@@ -28,20 +28,16 @@ func TestStartupCommands(t *testing.T) {
 
 	commands, err := StartupCommands(root)
 	require.NoError(t, err)
-	require.Len(t, commands, 3)
+	require.Len(t, commands, 2)
 
 	assert.Equal(t, PublishCommandKind, commands[0].Kind)
-	assert.Equal(t, "nest/units/controller_1/discovery", commands[0].Publish.Topic)
+	assert.Equal(t, "homeassistant/device/nest_controller_1_unit/config", commands[0].Publish.Topic)
 	assert.True(t, commands[0].Publish.Retain)
-	assert.Contains(t, string(commands[0].Publish.Payload), `"commands_enabled":false`)
+	assert.Contains(t, string(commands[0].Publish.Payload), `"cmps"`)
+	assert.Contains(t, string(commands[0].Publish.Payload), `"unique_id":"nest_controller_1_office_light"`)
+	assert.Contains(t, string(commands[0].Publish.Payload), `"command_topic":"nest/units/controller_1/lights/office_light/command"`)
 	assert.Equal(t, PublishCommandKind, commands[1].Kind)
-	assert.Equal(t, "homeassistant/device/nest_controller_1_unit/config", commands[1].Publish.Topic)
+	assert.Equal(t, "nest/units/controller_1/availability", commands[1].Publish.Topic)
+	assert.Equal(t, []byte("online"), commands[1].Publish.Payload)
 	assert.True(t, commands[1].Publish.Retain)
-	assert.Contains(t, string(commands[1].Publish.Payload), `"cmps"`)
-	assert.Contains(t, string(commands[1].Publish.Payload), `"unique_id":"nest_controller_1_office_light"`)
-	assert.Contains(t, string(commands[1].Publish.Payload), `"command_topic":"nest/units/controller_1/lights/office_light/command"`)
-	assert.Equal(t, PublishCommandKind, commands[2].Kind)
-	assert.Equal(t, "nest/units/controller_1/availability", commands[2].Publish.Topic)
-	assert.Equal(t, []byte("online"), commands[2].Publish.Payload)
-	assert.True(t, commands[2].Publish.Retain)
 }
